refactor(context): extract context selection module
